requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-huggingface
langchain-community
chromadb
sentence-transformers
pydantic